+++ NEWS.md
@@ -1,3 +1,11 @@
+# refundDevel
+
+* homogenized inputs/outputs to most fpca.XXX functions
+
+* fix documentation for fpca.face
+
+* added pco ridge regression see ?poridge
+
 # refund 0.1-14
 
 * add `fpca.lfda()` function